fix race condition with memory consumption by using ray init fixture
